+++ nope/lib/assets/js/ui.js
@@ -53,10 +53,10 @@
       }
     }])
     .filter('nopeMoment', [function() {
-      return function(input, format, timezone) {
+      return function(input, format, t) {
         input = input ? input.split(' ').join('T') + 'Z' : input;
         format = format || 'fromNow';
-        return moment(input, 'YYYY-MM-DD hh:mm:ss')[format]();
+        return moment(input, 'YYYY-MM-DD hh:mm:ss')[format](t);
       }
     }])
     .filter('nopeGetIds', [function() {
@@ -381,30 +381,9 @@
         restrict: 'E',
         replace: true,
         require: 'ngModel',
-        template: '<div>\
-          <ul dnd-list="ngModel" class="list-group list-group-contents is-multiple" ng-show="ngModel && preview" ng-if="multiple">\
-            <li class="list-group-item" ng-repeat="item in ngModel" dnd-draggable="item" dnd-moved="ngModel.splice($index,1)">\
-              <i class="fa fa-bars handle"></i>\
-              <img dnd-nodrag class="img-thumbnail preview" ng-src="{{item.preview[preview]}}" ng-if="hasPreview && item.preview[preview]" />\
-              <span dnd-nodrag class="title">{{item[titleField]}}</span>\
-              <div dnd-nodrag class="btn-group btn-group-xs toolbar">\
-                <a href="" class="btn" ng-click="ngModel.swapItems($index, $index-1);" ng-if="!$first"><i class="fa fa-arrow-up"></i></a>\
-                <a href="" class="btn" ng-click="ngModel.swapItems($index, $index+1);" ng-if="!$last"><i class="fa fa-arrow-down"></i></a>\
-                <a href="" class="btn text-danger" ng-click="ngModel.removeItemAt($index);"><i class="fa fa-times-circle"></i></a>\
-              </div>\
-            </li>\
-          </ul>\
-          <ul class="list-group list-group-contents" ng-show="ngModel && preview" ng-if="!multiple">\
-            <li class="list-group-item">\
-              <img class="img-thumbnail preview" ng-src="{{ngModel.preview[preview]}}" ng-if="hasPreview && ngModel.preview[preview]" />\
-              <span class="title">{{ngModel[titleField]}}</span>\
-              <div class="btn-group btn-group-xs toolbar pull-right">\
-                <a href="" class="btn text-danger" ng-click="remove();"><i class="fa fa-times-circle"></i></a>\
-              </div>\
-            </li>\
-          </ul>\
-          <a href="" class="btn btn-block btn-default" ng-click="openModal($event)" ng-hide="!multiple && ngModel">{{label || \'Add\'}} <i class="fa fa-plus"></i></a>\
-        </div>',
+        templateUrl: function($element, $attrs) {
+          return 'view/directive/model/'+($attrs.template || 'content')+'.html'
+        },
         scope: {
           multiple: '=?',
           ngModel: '=',
@@ -417,9 +396,8 @@
         controller: ['$scope', '$element', '$attrs', function($scope, $element, $attrs) {
           var theModal;
 
-          $scope.titleField = $scope.title || 'title';
-          $scope.multiple = (angular.isDefined($scope.multiple) ? $scope.multiple : true);
-          $scope.hasPreview = !!$scope.preview;
+          $scope.preview = $scope.preview || 'icon';
+          $scope.multiple = (angular.isDefined($scope.multiple) ? $scope.multiple : false);
 
           $scope.openModal = function($event) {
             $scope.selection = [];