+++ nope/lib/views/gallery/form.php
@@ -35,11 +35,11 @@
       </div>
       <div class="form-group">
         <label>Cover</label>
-        <nope-model href="#/media" ng-model="gallery.cover" multiple="false" label="Add gallery cover" preview="icon"></nope-model>
+        <nope-model href="#/media" ng-model="gallery.cover" multiple="false" label="Add gallery cover" preview="icon" template="media"></nope-model>
       </div>
       <div class="form-group">
         <label>Media</label>
-        <nope-model href="#/media?excluded={{(gallery.media | nopeGetIds).join(',')}}" ng-model="gallery.media" label="Add media" preview="icon"></nope-model>
+        <nope-model href="#/media?excluded={{(gallery.media | nopeGetIds).join(',')}}" ng-model="gallery.media" label="Add media" preview="icon" template="media"></nope-model>
       </div>
     </div>
     <div class="panel-footer">